requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.6
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
executing==1.2.0
fastapi==0.99.1
flatbuffers==23.5.26
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.268
langsmith==0.0.25
marshmallow==3.20.1
matplotlib-inline==0.1.6
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openai==0.27.8
overrides==7.4.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
posthog==3.0.2
prompt-toolkit==3.0.39
protobuf==4.24.1
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pydantic==1.10.12
Pygments==2.16.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.20
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.4.0
tokenizers==0.13.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
